+++ src/styles/_bootstrap-overrides.scss
@@ -14,7 +14,8 @@
 /* ======== Bootstrap Carousel style overrides ======== */
 .carousel {
     .carousel-indicators {
-        bottom: -50px;
+        bottom: -70px;
+        margin-bottom: 36px;
 
         .active {
             opacity: 1;
